requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
pytz==2024.2
requests==2.32.3
gunicorn==23.0.0
